third_party/blink/web_tests/inspector-protocol/page/set-intercept-file-chooser-dialog-expected.txt

Tests that Page.setInterceptFileChooserDialog works as expected

Running test: testAcceptFile
file chooser mode: selectSingle
selected files: ["path1"]

Running test: testAcceptMultipleFiles
file chooser mode: selectMultiple
selected files: ["path1","path2"]

Running test: testShowPickerAPI
file chooser mode: selectSingle
selected files: ["path1"]

Running test: testOpenFilePickerAPI
Failed to execute 'showOpenFilePicker' on 'Window': Intercepted by Page.setInterceptFileChooserDialog().

Running test: testSaveFilePickerAPI
Failed to execute 'showSaveFilePicker' on 'Window': Intercepted by Page.setInterceptFileChooserDialog().

Running test: testDirectoryPickerAPI
Failed to execute 'showDirectoryPicker' on 'Window': Intercepted by Page.setInterceptFileChooserDialog().

Running test: testErrors
Try enabling file interception in multiclient
{
    id : <number>
    result : {
    }
    sessionId : <string>
}
